feat(schema) new schema validation engine
A new schema validation engine, supporting various
data types and validations.

---

Adds a MetaSchema schema for validating schemas, to be used
especially when validating third-party plugin schemas.
Minor adjustments are made to the schema engine to support it.

---

Introduce a kong.db.schemas.typedefs module, storing common
type definitions. We use the name typedefs (and not types)
to make it explicity that these are type synonyms a la
typedefs in the C language, and not distinct types. In other
words, two type definitions that are structurally equal will
match identically as far as schema validation is concerned.

---

Turns the `fields` entry of a schema into an array, so that
each field is a one-key map. Adds a utility iterator,
`schema:each_field()` to make traversing fields easier.
Also updates tests accordingly.

---

Much to Tony Hoare's dismay, we introduce in Kong schemas the
billion-dollar mistake [1] : nullable types.

We do it in full understanding of how terrible an idea this is,
and in full appreciation that we are interfacing with the world
of JSON, JavaScript and the web, a world which expects null
references to exist and which has a rather quaint view of type
systems [2].

In order to keep the beast contained as much as possible, we
shy away from performing any automatic coercions (at this
level, at least -- handling form-encoded inputs elsewhere in
Kong presents challenges of its own).

The semantics introduced in this commit are as follows:

* All fields are nullable by default
* A field can be made non-nullable setting `nullable = false`
* Sub-values in compound fields (maps, arrays, sets and records)
  are NOT nullable.
* Foreign entries are subject to its own schema nullability
  settings.
* No value compares as equal to ngx.null.

The humorous nature of this commit message is brought to
you by the fact that it will be eventually lost to the sands
of time when we squash-commit this branch.

[1] https://www.infoq.com/presentations/Null-References-The-Billion-Dollar-Mistake-Tony-Hoare
[2] https://www.bram.us/wordpress/wp-content/uploads/2015/01/js-equality.png

---

Strings cannot be empty unless said otherwise using `len_min = 0`.
This changes the behavior to effectively get rid of empty strings,
while leaving the door open for any future use-cases that may
eventually need them.

---

Adds a `validate_update` function with proper semantics
for partial updates.

For symmetry, it also adds a `validate_insert` function, which
is equivalent to the current incarnation of `validate`.

Introduces `entity_checks` to schemas, which are declarative
checks designed to apply over multiple fields. Hopefully,
these will phase out `self.check`. Errors detected by
entity checks are available in `errors["@entity"]` and also
separately as a third argument.

Three entity checks are currently supported:

* `at_least_one_of` - at least one of the fields is given
* `only_one_of` - only one of the fields is given
* `conditional` - a case study of a more powerful entity
  check for testing the custom validator machinery: it
  works like "if/then": it tries a validation in one field,
  and if that passes, checks a validation in another field.
  Any one of the field validators can be composed with it.
  By capturing the notion of two-field interdependence,
  hopefully this will save us from creating a zillion
  similar validators. (And yes, I thought about adding an
  "else" mode -- will do it when the appropriate
  use-case arrives.)

With all these features in hand, I uncommented the partial
schema validaton from `update()` in the DAO, and adjusted some
tests accordingly.

Sorry about the long multi-topic commit, but due to the
interdependent nature of these features (no pun intended!)
it would be a lot of useless work to atomize this into multiple
commits in this branch. (I'd do it if these commits survived, though.)

---

An Entity is a restricted sub-type of Schema, which refers to
entities that can be persisted through the DAO.

An initial set of constraints is defined on Entity objects,
namely:

* no fields can be set to nil
* maps can only have strings as keys
* aggregate types only aggregate on basic types

These constraints can be later easily relaxed as we see fit
if we decide to give more power to DAO entities.

The general Schema feature set (including nil) is retained
as it is still useful for MetaSchema validation
(note that the MetaSchema is a more powerful sort of schema,
including for instance function fields).

The support for nil was renamed from `required = false` into
`nilable` (which is only available for the MetaSchema to use),
fixing the awkwardness of the fact that `required` was
actually a tri-state variable with different behaviors for
true, false and nil.

The Postgres strategy was adapted reflecting the change in the
routes schema.

---

feat(schema) allow sets to be indexed by their elements

When declaring a set, for set of strings,
you also get a my_set.my_entry shortcut syntax.
(We conservatively add this feature to strings only
because it cannot be applied to all types:
using this in sets of numbers would produce a conflict,
and some other types are not referentially transparent.)

---

feat(schema) `match_all`, `match_none`, `match_any` validators

Adds three new validators for lists of string patterns:
`match_all`, `match_none` and `match_any`.

This avoids introducing polymorphic validators to the schema engine,
making `match` and `not_match` operate on a single pattern only.

---

feat(schema) always run validators in the same order

Counting this as a feature rather than a bugfix, since
it improves on the specification.
This makes our tests more reliable.

---

feat(schema) hard-code validators ordering

Instead of relying on their alphabetical sort, we order validators per
the following rules:

1. Most "narrow"/limiting validators first
2. Negative validators
3. Positive validators
4. Custom validators

This is to avoid overriding error messages/validation of "simpler" rules
being overridden by rules validating larger scopes.

We also avoid building the ordering of validators for each call to
`validate()`.